get rid of DVExportNew, merging it into DVExport
